Git .lock file problem has been fixed, and Eric is now using git

+++ 1_mapping.py
@@ -8,7 +8,7 @@ to new reference sequences generated as the consensus of each SAM
 file.
 
 Dependencies:
-	pipeline2conseq_v2.py
+	pileup2conseq_v2.py
 	sam2fasta.py
 """
 
@@ -21,7 +21,7 @@ from sam2fasta import *
 
 
 
-def remap (f1, f2, samfile, ref):
+def remap (f1, f2, samfile, ref, qCutoff=30):
 	"""
 	Generate a sample-specific consensus sequence from a samtools
 	PILEUP file, and then remap everything to this consensus as a
@@ -29,32 +29,41 @@ def remap (f1, f2, samfile, ref):
 	consensus sequence file.
 	"""
 	bamfile = samfile.replace('.sam', '.bam')
+
+	# Convert sam to bam (fai file is a static file containing reference sequence)
 	os.system('samtools view -bt %s.fasta.fai %s > %s 2>/dev/null' % (ref, samfile, bamfile))
+
+	# Prerequisite for pileup - sort the bam (And needs to a bam for pileup generation)
 	os.system('samtools sort %s %s.sorted' % (bamfile, bamfile))
+
+	# Generate the pileup
 	os.system('samtools mpileup -A %s.sorted.bam > %s.pileup 2>/dev/null' % (bamfile, bamfile))
 	
 	# generate consensus sequence from pileup
-	os.system('python pileup2conseq_v2.py %s.pileup' % bamfile)
+	os.system("python pileup2conseq_v2.py {}.pileup {}".format(bamfile, qCutoff))
 	confile = bamfile+'.pileup.conseq'
 	os.system('bowtie2-build -q -f %s %s' % (confile, confile))
-	
-	# mapping against new reference
+
+	# The new file is named *.remap.sam
 	samfile = samfile.replace('.sam', '.remap.sam')
-	
-	os.system('bowtie2 --quiet -p 4 --local -x %s -1 %s -2 %s -S %s\
+
+	# Default behavior of bowtie2 expects fastq files (In this case, on R1 and R2)
+	os.system('bowtie2 --quiet -p 1 --local -x %s -1 %s -2 %s -S %s\
 				--no-unal --met-file %s --un %s --un-conc %s' % (confile, 
 			f1, f2, samfile, samfile.replace('.sam', '.bt2_metrics'),
 			samfile.replace('.sam', '.bt2_unpaired_noalign.fastq'), 
 			samfile.replace('.sam', '.bt2_paired_noalign.fastq')))
 	
+	# -f for fasta
+
 	return samfile, confile
 
-
+refpath = sys.argv[1]
+f = sys.argv[2]
+qCutoff = int(sys.argv[3])
 
 
 # Get names of references
-refpath = sys.argv[1]
-
 infile = open(refpath+'.fasta', 'rU')
 refnames = []
 for line in infile:
@@ -62,10 +71,7 @@ for line in infile:
 		refnames.append(line.strip('>\n'))
 infile.close()
 
-
-
 # path to R1 FASTQ
-f = sys.argv[2]
 root = '/'.join(f.split('/')[:-1])
 filename = f.split('/')[-1]
 prefix = filename.split('.')[0].split('_')[0]
@@ -73,10 +79,9 @@ prefix = filename.split('.')[0].split('_')[0]
 f1 = f
 f2 = f.replace('R1', 'R2')
 
-
 # Initial mapping to construct reference
 samfile = '%s/%s.prelim.sam' % (root, prefix)
-os.system('bowtie2 --quiet -p 4 --local -x %s -1 %s -2 %s -S %s' % (refpath, 
+os.system('bowtie2 --quiet -p 1 --local -x %s -1 %s -2 %s -S %s' % (refpath, 
 		f1, 
 		f2, 
 		samfile))
@@ -119,11 +124,17 @@ for refname in refnames:
 		continue
 	
 	samfile = refsams[refname]['handle'].name
-	samfile, confile = remap(f1, f2, samfile, refpath)
+	samfile, confile = remap(f1, f2, samfile, refpath, qCutoff)
+
+# just for now, keep all the files around!
+sys.exit()
 
 # clean up intermediate BAM files
 bamfiles = glob(root+'/'+prefix+'.*bam*')
 for bamfile in bamfiles:
+	if bamfile.endswith('.conseq'):
+		# spare the consensus sequence
+		continue
 	os.remove(bamfile)
 
 # clean up intermediate SAM files
